fix(mobile,supabase): resolve PHI on cold start, dedupe offline preset user id, reuse parseProfileAppRole
