+++ apps/mobile/src/lib/symptom-presets/symptom-preset-service.ts
@@ -23,9 +23,8 @@ import {
   updatePresetSymptom,
   updateSymptomPreset,
 } from '@abstrack/supabase';
-import type { PowerSyncDatabase } from '@powersync/react-native';
 import { fetchMobileDeviceIsConnected } from '../network/mobile-device-netinfo';
-import { resolveMobilePhiSubjectUserContext } from '../phi-subject/resolve-mobile-phi-subject-user-context';
+import { resolveOfflinePresetListScopeUserId } from '../phi-subject/resolve-offline-preset-list-scope-user-id';
 import { listSymptomPresetsForUserFromPowerSyncDb } from '../powersync/powersync-episode-flow-reads';
 import {
   clarifyNetworkErrorWhenReplicaUnavailable,
@@ -33,84 +32,9 @@ import {
   resolvePowerSyncDatabaseForOfflineRead,
   type PowerSyncOfflineReadContext,
 } from '../powersync/powersync-offline-read-bridge-snapshot';
-import {
-  getMobileAuthSessionSafe,
-  getMobileSupabaseClient,
-  isAuthSessionRecoveryFailure,
-  readPersistedMobileAuthUserId,
-} from '../supabase-wiring';
+import { getMobileSupabaseClient } from '../supabase-wiring';
 
-/**
- * Resolves the signed-in user id from the persisted session (same pattern as episode templates).
- * Uses {@link getMobileAuthSessionSafe} rather than `getUser()` so airplane mode does not fail the auth lookup.
- * When that helper returns `auth_session_recovery_failed`, falls back to {@link readPersistedMobileAuthUserId}
- * so transient secure-store / recovery hiccups do not block offline PowerSync preset lists.
- *
- * @returns `{ ok: true, data: id }` when signed in; `{ ok: true, data: null }` when signed out with
- * no auth error; `{ ok: false, error }` when the session read failed.
- */
-export async function getCurrentUserId(): Promise<
-  PresetDataResult<string | null>
-> {
-  try {
-    const {
-      data: { session },
-      error,
-    } = await getMobileAuthSessionSafe();
-    if (!error) {
-      return { ok: true, data: session?.user?.id ?? null };
-    }
-    if (!isAuthSessionRecoveryFailure(error)) {
-      return { ok: false, error: toPresetDataError(error) };
-    }
-    const persistedId = await readPersistedMobileAuthUserId();
-    if (persistedId != null) {
-      return { ok: true, data: persistedId };
-    }
-    return { ok: false, error: toPresetDataError(error) };
-  } catch (caught) {
-    return { ok: false, error: toPresetDataError(caught) };
-  }
-}
-
-/**
- * Resolves which `user_id` to filter on for offline replica preset lists.
- *
- * When `explicitScopeUserId` is a non-empty string, returns it. Otherwise, when `replicaDb` is
- * set, uses {@link resolveMobilePhiSubjectUserContext} so caretaker sessions query the linked
- * patient’s rows instead of the caretaker auth uid. Falls back to {@link getCurrentUserId} when no
- * replica is available or PHI scope is unavailable.
- *
- * @param explicitScopeUserId - Optional PHI row owner from the caller (skips resolver when set).
- * @param replicaDb - Open PowerSync database when reading SQLite; enables PHI resolution when scope is omitted.
- */
-async function resolveOfflinePresetListScopeUserId(
-  explicitScopeUserId: string | null | undefined,
-  replicaDb: PowerSyncDatabase | null,
-): Promise<PresetDataResult<string | null>> {
-  if (explicitScopeUserId != null) {
-    const trimmed = explicitScopeUserId.trim();
-    if (trimmed !== '') {
-      return { ok: true, data: trimmed };
-    }
-  }
-  if (replicaDb) {
-    const phi = await resolveMobilePhiSubjectUserContext({
-      powerSyncDatabase: replicaDb,
-    });
-    if (!phi.ok) {
-      return phi;
-    }
-    const subject =
-      phi.data?.phiSubjectUserId != null
-        ? phi.data.phiSubjectUserId.trim()
-        : '';
-    if (subject !== '') {
-      return { ok: true, data: subject };
-    }
-  }
-  return getCurrentUserId();
-}
+export { getMobileAuthUserIdForPresetListOffline as getCurrentUserId } from '../phi-subject/resolve-offline-preset-list-scope-user-id';
 
 /**
  * Lists the signed-in user’s symptom presets, falling back to the PowerSync replica when Supabase
@@ -126,7 +50,7 @@ async function resolveOfflinePresetListScopeUserId(
  * @param options.powerSyncOfflineRead - Prefer passing `usePowerSyncBridgeState()` fields from the
  *   screen so reads use the same DB instance as `PowerSyncContext` (PowerSync SDK lifecycle).
  * @param options.scopeUserId - Optional PHI row owner (`user_id`) for replica SQL. When omitted but
- *   a replica handle is used, {@link resolveMobilePhiSubjectUserContext} supplies the subject
+ *   a replica handle is used, {@link resolveOfflinePresetListScopeUserId} supplies the subject
  *   (caretaker → linked patient). Otherwise {@link getCurrentUserId} is used.
  * @returns {@link PresetDataResult} of preset rows or an error.
  */
